+++ src/handler/svelte.ts
@@ -1,6 +1,6 @@
 import fs from "fs";
 import path from "path";
-import { select, cancel, text } from "@clack/prompts";
+import { select, cancel, text, confirm } from "@clack/prompts";
 
 const fsPromises = fs.promises;
 
@@ -32,6 +32,7 @@ export async function handler() {
     const templatePath = path.join(
       __dirname,
       "..",
+      "..",
       "templates",
       "svelte",
       "component.ts"
@@ -44,7 +45,19 @@ export async function handler() {
       String(componentName)
     );
 
-    await fsPromises.writeFile(filePath, replacedData);
+    if (fs.existsSync(filePath)) {
+      const shouldContinue = await confirm({
+        message:
+          "The component you are trying to create already exists, do you want to continue?",
+      });
+      if (shouldContinue) {
+        await fsPromises.writeFile(filePath, replacedData);
+      }
+      if (!shouldContinue) {
+        cancel("Operation cancelled");
+        process.exit(0);
+      }
+    }
 
     console.log(
       `Successfully created the component named ${String(componentName)}!`
